Fix condition where message was broadcast after removing the channel
The disconnect message was being sent after the channel no longer
existed, causing getChannelByName to fail. This was happening when a
client uncleanly disconnected. I think it should be OK for disconnect to
send the disconnect and the modify the mvar nonatomically, because no
more messages will be received from the disconnecting client.